This commit adds a dropdown to the GUI that allows users to switch the hardware trigger line. The dropdown is populated with the available trigger lines, and the selected line is passed to the camera object. The camera object is then used to set the trigger line.
Changes were made to the logic of when the user can select a new trigger line:
- The user cannot select a new trigger line when hardware acquisition is in progress.
- The user cannot select a new trigger line when the camera is recording and the recording is being processed.

Fixed the order of the trigger line source console log output:
It was moved from the beginning of *stop hardware acquisition* to the end of *start hardware acquisition*.

---

- The physical hardware trigger line being switched has **not** yet been tested.
- Tested the GUI changes and the logic of when the user can select a new trigger line.
- The dropdown is populated with the available trigger lines, and the selected line is passed to the camera object.

---

Run the following once you are in the project directory:

```shell
source venv/bin/activate
python3 main_gui.pyw

```

- May have threading conflicts if buttons and dropdowns are selected in quick succession or in certain orders.

- The physical hardware trigger line switching has not been tested yet.

- Other buttons being selected in quick succession during any combination of button selections may cause issues.

- This is before the calibration and projection branches were merged.
- This is not merged with main.

This commit adds a dropdown to the GUI that allows users to switch the hardware trigger line.
The dropdown is populated with the available trigger lines, and the selected line is passed to the camera object. The camera object is then used to set the trigger line.

Changes were made to the logic of when the user can select a new trigger line:
- The user cannot select a new trigger line when hardware acquisition is in progress.
- The user cannot select a new trigger line when the camera is recording and the recording is being processed.

Fixed the order of the trigger line source console log output:
It was moved from the beginning of *stop hardware acquisition* to the end of *start hardware acquisition*.

---

- The physical hardware trigger line being switched has **not** yet been tested.
- Tested the GUI changes and the logic of when the user can select a new trigger line.
- The dropdown is populated with the available trigger lines, and the selected line is passed to the camera object.

---

Run the following once you are in the project directory:

```shell
source venv/bin/activate
python3 main_gui.pyw

```

- May have threading conflicts if buttons and dropdowns are selected in quick succession or in certain orders.

- The physical hardware trigger line switching has not been tested yet.

- Other buttons being selected in quick succession during any combination of button selections may cause issues.

- This is before the calibration and projection branches were merged.
- This is not merged with main.